+++ stories/dnd/dnd.stories.js
@@ -1,60 +1,13 @@
-import React, { useRef, useState, useEffect, useMemo, useCallback } from 'react';
-import makeDroppable from '../../src/components/DragAndDrop/makeDroppable';
-import makeDraggable from '../../src/components/DragAndDrop/makeDraggable';
+import React, { useMemo, useRef } from 'react';
 import '../../src/styles/_all.scss';
-import { DnDProvider, useDnDContext } from '../../src/components/DragAndDrop/DnDContext';
-import DragManager from '../../src/components/DragAndDrop/DragManager';
-import DraggablePreview from '../../src/components/DragAndDrop/DragPreview';
-import useResizeObserver from 'use-resize-observer';
-import v4 from 'uuid/v4';
+import { DnDProvider } from '../../src/components/DragAndDrop/DnDContext';
 import DRAGGABLE_TYPES from '../../src/components/DragAndDrop/draggableTypes';
+import useDroppable from '../../src/components/DragAndDrop/useDroppable';
+import useDraggable from '../../src/components/DragAndDrop/useDraggable';
 
 export default { title: 'Systems/DND' };
 
-const useDroppable = (
-  componentRef,
-  typesAccepted,
-  dropHandler,
-) => {
-  const { width, height } = useResizeObserver({ ref: componentRef });
-
-  const uuid = useMemo(() => v4(), []);
-
-  const {
-    upsertTarget,
-    removeTarget,
-    getActiveState,
-  } = useDnDContext();
-
-  const isActive = getActiveState(uuid)();
-
-  const updateTarget = () => {
-    console.log('updateTarget');
-    if (!componentRef || !(height && width)) { return; }
-
-    const { x, y } = componentRef.current.getBoundingClientRect();
-
-    upsertTarget({
-      uuid,
-      accepts: typesAccepted,
-      width,
-      height,
-      y,
-      x,
-    });
-  };
-
-  useEffect(() => {
-    updateTarget();
-    return () => removeTarget(uuid);
-  }, [width, height]);
-
-  return {
-    isActive,
-  };
-};
-
-const TestDiv = ({
+const TestDroppable = ({
   accepts = [DRAGGABLE_TYPES.CARD, DRAGGABLE_TYPES.NODE],
   children
 }) => {
@@ -89,119 +42,7 @@ const TestDiv = ({
   );
 };
 
-const useDraggable = (
-  componentRef,
-  previewRef,
-  type,
-  payload,
-  previewComponent
-) => {
-  const [isDragging, setIsDragging] = useState(false);
-
-  let dragManager = null;
-  let previewEl = null;
-
-  const {
-    setSource,
-    overObstable,
-    overDropTarget,
-  } = useDnDContext();
-
-  const cleanupDragManager = () => {
-    if (dragManager) {
-      dragManager.unmount();
-      dragManager = null;
-    }
-  };
-
-  const cleanupPreview = () => {
-    if (previewEl) {
-      previewEl.cleanup();
-      previewEl = null;
-    }
-  };
-
-  const createPreview = () => {
-    // if (!previewComponent) {
-    //   previewEl = new DraggablePreview(componentRef.current);
-    //   return;
-    // }
-
-    previewEl = new DraggablePreview(componentRef.current);
-  };
-
-  const updatePreview = ({ x, y }) => {
-    if (previewEl) {
-      previewEl.position({ x, y });
-    }
-  };
-
-  const onDragStart = (movement) => {
-    createPreview();
-
-    console.log('onDragStart', movement);
-
-    setSource({
-      type,
-    });
-
-    setIsDragging(true);
-  };
-
-  // const throttledDragAction = throttle(({ x, y, ...other }) => {
-  //   console.log('Drag',x,y);
-  //   // store.dispatch(
-  //   //   actions.dragMove({
-  //   //     x, y, setValidMove, ...other,
-  //   //   }),
-  //   // );
-  // }, 60);
-
-  const onDragMove = ({ x, y, ...other }) => {
-    updatePreview({ x, y });
-    // throttledDragAction({ x, y, ...other });
-  };
-
-  const onDragEnd = (movement) => {
-    cleanupPreview();
-    setIsDragging(false);
-
-    setSource(null);
-
-    // store.dispatch(
-    //   actions.dragEnd(movement),
-    // );
-  };
-
-  const styles = () => (isDragging ? { visibility: 'hidden' } : {});
-
-  useEffect(() => {
-    if (componentRef.current) {
-      dragManager = new DragManager({
-        el: componentRef.current,
-        onDragStart,
-        onDragMove,
-        onDragEnd,
-        // scrollDirection,
-      });
-    }
-
-    return () => {
-      cleanupPreview();
-      cleanupDragManager();
-    };
-  }, [componentRef]);
-
-
-
-  return {
-    styles: styles(),
-    positionIsValid: getPositionIsValid(),
-    positionIsDroppable: getPositionIsDroppable(),
-  };
-};
-
-const TestCard = (props) => {
+const TestDraggable = (props) => {
   const ref = useRef(null);
   const previewRef = useRef(null);
 
@@ -216,6 +57,7 @@ const TestCard = (props) => {
     styles,
     positionIsValid, // Changes when hovering obstacles
     positionIsDroppable, // Changes when hovering a valid drop target
+    isDragging, // True when this item is being dragged
   } = useDraggable(
     ref,
     previewRef,
@@ -224,13 +66,23 @@ const TestCard = (props) => {
     previewComponent,
   );
 
+  console.log({ isDragging, positionIsValid});
+
+  const opacity = useMemo(() => {
+    if (!isDragging) {
+      return 1;
+    }
+
+    return positionIsValid ? 1 : 0;
+  }, [isDragging, positionIsValid]);
+
   return (
     <div
       ref={ref}
       style={{
         ...styles,
         background: positionIsDroppable ? 'blue' : 'orange',
-        opacity: positionIsValid ? 1 : 0,
+        opacity,
         width: '5rem',
         height: '5rem',
       }}
@@ -246,12 +98,12 @@ const Template = (args) => {
   return (
     <>
       <DnDProvider>
-        <TestDiv
+        <TestDroppable
           accepts={DRAGGABLE_TYPES.CARD}
           // activePlaceholder={}
           // hoverPlaceholder={}
         />
-        <TestDiv
+        <TestDroppable
           accepts={DRAGGABLE_TYPES.NODE}
           // activePlaceholder={}
           // hoverPlaceholder={}
@@ -262,11 +114,11 @@ const Template = (args) => {
             height: '15rem',
           }}
         >
-          <TestCard />
-          <TestCard />
-          <TestCard />
-          <TestCard />
-          <TestCard />
+          <TestDraggable />
+          <TestDraggable />
+          <TestDraggable />
+          <TestDraggable />
+          <TestDraggable />
         </div>
 
       </DnDProvider>
